+++ CHANGELOG.md
@@ -29,6 +29,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Tests for "The stonemason".
 - Tests for "Continued fractions".
 - Tests for "All operations are equal!".
+- Tests for "Battle tower".
 
 ## [1.14.0] - 2022-10-31
 ### Added